feat: Add nginx container
Adds the nginx container, which is going to proxy our requests to both `client` and `server` containers.
A renaming of the `server` container to `api` was required, because `server` is a reserved word when writing the nginx configuration file.
The nginx will proxy the `/` requests to the `client` container, and `/api` requests to the `api` container.